putting in an enum file the enum buttons
