+++ src/enums/buttons.ts
@@ -1,0 +1,19 @@
+export enum Button {
+	UP,
+	DOWN,
+	LEFT,
+	RIGHT,
+	SUBMIT,
+	ACTION,
+	CANCEL,
+	MENU,
+	STATS,
+	CYCLE_SHINY,
+	CYCLE_FORM,
+	CYCLE_GENDER,
+	CYCLE_ABILITY,
+	CYCLE_NATURE,
+	CYCLE_VARIANT,
+	SPEED_UP,
+	SLOW_DOWN
+}
