+++ src/ui/command-ui-handler.ts
@@ -5,7 +5,7 @@ import PartyUiHandler, { PartyUiMode } from "./party-ui-handler";
 import { Mode } from "./ui";
 import UiHandler from "./ui-handler";
 import i18next from '../plugins/i18n';
-import {Button} from "../inputs-controller";
+import {Button} from "../enums/buttons";
 
 export enum Command {
   FIGHT = 0,
