+++ src/ui/egg-hatch-scene-handler.ts
@@ -2,7 +2,7 @@ import BattleScene from "../battle-scene";
 import { EggHatchPhase } from "../egg-hatch-phase";
 import { Mode } from "./ui";
 import UiHandler from "./ui-handler";
-import {Button} from "../inputs-controller";
+import {Button} from "../enums/buttons";
 
 export default class EggHatchSceneHandler extends UiHandler {
   public eggHatchContainer: Phaser.GameObjects.Container;
