+++ src/ui/evolution-scene-handler.ts
@@ -2,7 +2,7 @@ import BattleScene from "../battle-scene";
 import MessageUiHandler from "./message-ui-handler";
 import { TextStyle, addTextObject } from "./text";
 import { Mode } from "./ui";
-import {Button} from "../inputs-controller";
+import {Button} from "../enums/buttons";
 
 export default class EvolutionSceneHandler extends MessageUiHandler {
   public evolutionContainer: Phaser.GameObjects.Container;
