+++ src/ui/ui.ts
@@ -35,7 +35,7 @@ import SavingIconHandler from './saving-icon-handler';
 import UnavailableModalUiHandler from './unavailable-modal-ui-handler';
 import OutdatedModalUiHandler from './outdated-modal-ui-handler';
 import SessionReloadModalUiHandler from './session-reload-modal-ui-handler';
-import {Button} from "../inputs-controller";
+import {Button} from "../enums/buttons";
 
 export enum Mode {
   MESSAGE,
